+++ src/main/java/gov/ca/cwds/idm/service/validation/ValidationService.java
@@ -1,0 +1,16 @@
+package gov.ca.cwds.idm.service.validation;
+
+import com.amazonaws.services.cognitoidp.model.UserType;
+import gov.ca.cwds.UniversalUserToken;
+import gov.ca.cwds.idm.dto.User;
+import gov.ca.cwds.idm.dto.UserUpdate;
+
+public interface ValidationService {
+
+  void validateUserCreate(UniversalUserToken admin, User user);
+
+  void validateUpdateUser(UniversalUserToken admin, UserType existedCognitoUser, UserUpdate updateUserDto);
+
+  boolean isActiveRacfIdPresent(String racfId);
+
+}
