+++ src/test/java/gov/ca/cwds/idm/service/TestHelper.java
@@ -3,13 +3,15 @@ package gov.ca.cwds.idm.service;
 import static gov.ca.cwds.util.UniversalUserTokenDeserializer.ADMIN_OFFICE_IDS_PARAM;
 import static gov.ca.cwds.util.UniversalUserTokenDeserializer.COUNTY_NAME_PARAM;
 
+import com.amazonaws.services.cognitoidp.model.UserType;
 import gov.ca.cwds.UniversalUserToken;
 import gov.ca.cwds.idm.dto.User;
+import gov.ca.cwds.idm.service.cognito.util.CognitoUtils;
 import java.util.Set;
 
-public final class AuthorizationTestHelper {
+public final class TestHelper {
 
-  private AuthorizationTestHelper() {
+  private TestHelper() {
   }
 
   public static User user(String countyName, String officeId) {
@@ -23,6 +25,25 @@ public final class AuthorizationTestHelper {
     User user = user(countyName, officeId);
     user.setRoles(roles);
     return user;
+  }
+
+  public static User user() {
+    User user = new User();
+    user.setEmail("gonzales@gmail.com");
+    user.setFirstName("Garcia");
+    user.setLastName("Gonzales");
+    user.setCountyName("Yolo");
+    user.setEnabled(Boolean.TRUE);
+    return user;
+  }
+
+  public static UserType userType(User user, String userId) {
+    UserType userType = new UserType();
+    userType.setUsername(userId);
+    userType.setEnabled(true);
+    userType.setUserStatus("FORCE_CHANGE_PASSWORD");
+    userType.withAttributes(CognitoUtils.buildCreateUserAttributes(user));
+    return userType;
   }
 
   static User withRole(String role) {
